update tentacat dependency
Prior to this commit gute_taten was pinned to a sha of Tentacat that did
not support paging through a user's repositories and events. This commit
bumps that dependency version so that *all* repos are examined and *all*
user events from at most 90 days or 300 events (as limited by the
GitHub API) are examined.

Close #1
Close #2